order-service/requirements.txt
Django==4.2.24
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-decouple==3.8
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
pytest==8.4.2
pytest-django==4.11.1
Faker==37.6.0
requests==2.32.3